프로토콜 정의와 선언 분리 + 접속시 세션 제공 (Separate definition and declaration of Protocol each other + give session when connect)

+++ Server/MainServer.cpp
@@ -36,7 +36,7 @@ bool MainServer::Run()
 	//serverAddr 셋팅
 	serverAddr.sin6_family = AF_INET6;
 	serverAddr.sin6_flowinfo = 0;
-	serverAddr.sin6_port = htons(port);
+	serverAddr.sin6_port = htons((uint16_t)port);
 	serverAddr.sin6_addr = in6addr_any;
 	//serverAddr.sin6_scope_id = if_nametoindex("etho");
